move: Product 클래스를 domain 계층으로 이동
