+++ bin/domain/entities/product.dart
